# feat(performance): HAVING Phase 4.3 - Performance optimization and benchmarking